order_tracking/requirements.txt
accept-paymob==1.0.0
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.3.4
astroid==2.7.3
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
bleach==4.1.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
CurrencyConverter==0.16.3
decorator==5.0.9
defusedxml==0.7.1
dicttoxml==1.7.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.4
django-ajax-selects==2.1.0
django-allauth==0.44.0
django-appconf==1.0.4
django-autoslug==1.9.8
django-bootstrap-datepicker-plus==3.0.5
django-cities==0.6
django-cities-light==3.8.1
django-ckeditor==6.1.0
django-countries==7.2.1
django-crispy==1.8.1
django-crispy-forms==1.12.0
django-decouple==2.1
django-filter==2.4.0
django-geojson==3.2.0
django-geoposition==0.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-leaflet==0.28.1
django-location-field==2.1.0
django-map-widgets==0.3.2
django-mapbox-location-field==1.7.3
django-model-utils==4.1.1
django-multiselectfield==0.1.12
django-rosetta==0.9.7
django-select2==7.7.1
django-tinymce==3.3.0
djangorestframework==3.12.4
docutils==0.17.1
filelock==3.0.12
future==0.18.2
GDAL @ file:///C:/Users/---Developer---/GDAL-3.3.2-cp38-cp38-win_amd64.whl
geocoder==1.38.1
geographiclib==1.52
geoip2==4.2.0
geopy==2.2.0
google-currency==1.0.9
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.8.1
ip2geotools==0.1.6
IP2Location==8.6.2
isort==5.9.3
keyring==23.1.0
lazy-object-proxy==1.6.0
lxml==4.6.3
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
oauthlib==3.1.1
packaging==21.0
paymob-accept==0.1.1
paypal==1.2.5
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
paypalrestsdk==1.13.1
Pillow==8.2.0
pip-review==1.1.0
pkginfo==1.7.1
platformdirs==2.3.0
polib==1.1.1
progressbar2==3.53.1
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyCurrency-Converter==1.0
Pygments==2.10.0
PyJWT==1.7.1
pylint==2.10.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyquery==1.4.3
python-utils==2.5.6
python3-openid==3.2.0
pytz==2021.1
pyuca==1.2
pywin32-ctypes==0.2.0
ratelim==0.1.6
readme-renderer==29.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
tqdm==4.61.2
twilio==6.60.0
twine==3.4.2
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.10.0.2
Unidecode==1.2.0
urllib3==1.26.5
virtualenv==20.4.7
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.3
zipp==3.5.0
